Move card and deeplink into their own files, add created mockio links and their expected JSON.
