+++ QRCodeTarot/ViewController.swift
@@ -24,8 +24,7 @@ class ViewController: UIViewController, QRCodeReaderViewControllerDelegate {
 
     func reader(_ reader: QRCodeReaderViewController, didScanResult result: QRCodeReaderResult) {
         result.value.url?.callCodable { (deepLink: DeepLink?) in
-            let card = deepLink?.applinks.details.first?.paths.first
-            print(card as Any)
+            print(deepLink?.card as Any)
         }
         dismiss(animated: true)
     }
@@ -48,37 +47,4 @@ class ViewController: UIViewController, QRCodeReaderViewControllerDelegate {
         super.viewDidAppear(animated)
         present(readerVC, animated: true)
     }
-}
-
-extension Card {
-
-    enum Symbol: String, Codable {
-        case hearts
-        case clubs
-        case spades
-        case diamonds
-    }
-}
-
-struct Card: Codable {
-    let symbol: Symbol
-    let number: Int
-    let deckVersion: Int
-}
-
-// MARK: - DeepLink
-struct DeepLink: Codable {
-    let applinks: Applinks
-}
-
-// MARK: - Applinks
-struct Applinks: Codable {
-    let details: [Detail]
-}
-
-// https://mocki.io/v1/74646ca2-9bce-4ad2-b623-f65f09cb3ac4 <- This can work as a QR code to show you with the internal scanner.
-// MARK: - Detail
-struct Detail: Codable {
-    let appID: String
-    let paths: [Card]
 }
